Split runtime and image service in different modules
I think it makes sense to split both implementations into separate
modules to separate the types and reduce the file size.
